+++ src/grscheller/boring_math/cli/ackermann.py
@@ -12,7 +12,7 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-"""Entry points for CLI scripts to evaluate Ackermann's function
+"""Entry points for CLI scripts to evaluate Ackermann's function.
 
 The version of Ackermann's function being used is defined recursively by